projects/009-jwt-authentication/gabrielegabellone/requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.7.0
typing_extensions==4.6.3
Werkzeug==2.3.5
